Move testing code to its own file
